Unet/requirements.txt
accelerate==0.31.0
appdirs==1.4.4
baal==1.9.2
Brotli==1.0.9
charset-normalizer==2.0.12
click==8.1.6
contourpy==1.1.0
cycler==0.11.0
diffusers==0.29.2
docker-pycreds==0.4.0
filelock==3.12.2
fonttools==4.40.0
fsspec==2024.5.0
gitdb==4.0.10
GitPython==3.1.36
gmpy2==2.1.5
h5py==3.11.0
huggingface-hub==0.23.4
idna==3.3
importlib-resources==5.12.0
importlib_metadata==8.0.0
Jinja2==3.1.2
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
networkx==2.6.3
numpy==1.26.4
opencv-python==4.7.0.72
packaging==23.1
pathtools==0.1.2
pillow==9.5.0
ply==3.11
protobuf==4.21.12
psutil==5.9.0
pyparsing==3.0.9
PyQt5==5.15.10
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.8.2
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentry-sdk==1.5.11
setproctitle==1.3.0
sip==6.7.11
six==1.16.0
smmap==3.0.5
structlog==22.3.0
sympy==1.12.1
threadpoolctl==3.5.0
tomli==2.0.1
torch==2.3.0
torchaudio==2.3.0
torchmetrics==0.9.3
torchvision==0.18.0
tornado==6.3.3
tqdm==4.64.1
triton==2.3.0
typing_extensions==4.8.0
unicodedata2==15.0.0
urllib3==2.0.3
wandb==0.15.5
zipp==3.17.0